Move PubSubService to lib/util
